# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CollectionBuildResponse": {
        "description": "确认构建后的批量投递结果，允许部分文献投递失败。",
        "properties": {
          "collection_id": {
            "format": "uuid",
            "title": "Collection Id",
            "type": "string"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/CollectionBuildRunResponse"
            },
            "title": "Runs",
            "type": "array"
          },
          "workflow_stage": {
            "title": "Workflow Stage",
            "type": "string"
          }
        },
        "required": ["collection_id", "workflow_stage", "runs"],
        "title": "CollectionBuildResponse",
        "type": "object"
      },
      "CollectionBuildRunResponse": {
        "description": "一次确认构建中单个入库运行的投递结果。",
        "properties": {
          "arq_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arq Job Id"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "ingestion_run_id": {
            "format": "uuid",
            "title": "Ingestion Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/IngestionRunStatus"
          }
        },
        "required": ["ingestion_run_id", "status"],
        "title": "CollectionBuildRunResponse",
        "type": "object"
      },
      "CollectionDocumentRemovalResponse": {
        "description": "移出待确认文献后的审计结果；文件不会被立即物理删除。",
        "properties": {
          "collection_paper_status": {
            "title": "Collection Paper Status",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "ingestion_run_status": {
            "$ref": "#/components/schemas/IngestionRunStatus"
          }
        },
        "required": ["document_id", "collection_paper_status", "ingestion_run_status"],
        "title": "CollectionDocumentRemovalResponse",
        "type": "object"
      },
      "CollectionDocumentResponse": {
        "description": "工作区中一篇活动文献及其最新入库运行。",
        "properties": {
          "access_rights": {
            "title": "Access Rights",
            "type": "string"
          },
          "added_at": {
            "format": "date-time",
            "title": "Added At",
            "type": "string"
          },
          "authors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Authors",
            "type": "array"
          },
          "byte_size": {
            "title": "Byte Size",
            "type": "integer"
          },
          "citation_text": {
            "title": "Citation Text",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "doi": {
            "title": "Doi",
            "type": "string"
          },
          "latest_ingestion_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IngestionRunResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "original_filename": {
            "title": "Original Filename",
            "type": "string"
          },
          "paper_id": {
            "format": "uuid",
            "title": "Paper Id",
            "type": "string"
          },
          "publication_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication Year"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "venue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue"
          }
        },
        "required": [
          "document_id",
          "paper_id",
          "doi",
          "title",
          "authors",
          "publication_year",
          "venue",
          "citation_text",
          "tags",
          "note",
          "original_filename",
          "byte_size",
          "source_url",
          "access_rights",
          "added_at",
          "latest_ingestion_run"
        ],
        "title": "CollectionDocumentResponse",
        "type": "object"
      },
      "CollectionDocumentsResponse": {
        "description": "活动文献列表与用于刷新页面的入库汇总。",
        "properties": {
          "collection_id": {
            "format": "uuid",
            "title": "Collection Id",
            "type": "string"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/CollectionDocumentResponse"
            },
            "title": "Documents",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/CollectionIngestionSummary"
          }
        },
        "required": ["collection_id", "documents", "summary"],
        "title": "CollectionDocumentsResponse",
        "type": "object"
      },
      "CollectionIngestionSummary": {
        "description": "集合页面展示的任务数量与可问答文献数量。",
        "properties": {
          "active_document_count": {
            "title": "Active Document Count",
            "type": "integer"
          },
          "ingestion_status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/IngestionRunStatus"
            },
            "title": "Ingestion Status Counts",
            "type": "object"
          },
          "researchable_document_count": {
            "title": "Researchable Document Count",
            "type": "integer"
          }
        },
        "required": ["active_document_count", "researchable_document_count"],
        "title": "CollectionIngestionSummary",
        "type": "object"
      },
...
      "CreateWorkspaceRequest": {
        "description": "新建研究工作区的最小输入，不允许客户端指定所有者或状态。",
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": ["name"],
        "title": "CreateWorkspaceRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IngestionRunResponse": {
        "description": "单次入库运行的持久化状态，适合列表轮询和失败诊断。",
        "properties": {
          "arq_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arq Job Id"
          },
          "attempt_no": {
            "title": "Attempt No",
            "type": "integer"
          },
          "chunking_config": {
            "additionalProperties": true,
            "title": "Chunking Config",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "embedding_config": {
            "additionalProperties": true,
            "title": "Embedding Config",
            "type": "object"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_current": {
            "title": "Is Current",
            "type": "boolean"
          },
          "parser_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Name"
          },
          "parser_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Version"
          },
          "pipeline_version": {
            "title": "Pipeline Version",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "statistics": {
            "additionalProperties": true,
            "title": "Statistics",
            "type": "object"
          },
          "status": {
            "$ref": "#/components/schemas/IngestionRunStatus"
          },
          "submitted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          }
        },
        "required": [
          "id",
          "document_id",
          "arq_job_id",
          "pipeline_version",
          "status",
          "stage",
          "parser_name",
          "parser_version",
          "chunking_config",
          "embedding_config",
          "statistics",
          "error_code",
          "error_message",
          "attempt_no",
          "is_current",
          "started_at",
          "finished_at",
          "created_at"
        ],
        "title": "IngestionRunResponse",
        "type": "object"
      },
      "IngestionRunStatus": {
        "description": "入库运行写入数据库的稳定状态值。",
        "enum": ["pending", "queued", "running", "completed", "failed", "cancelled"],
        "title": "IngestionRunStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkflowStageDisplay": {
        "description": "工作区研究阶段的中文展示元数据，不作为客户端提交输入。",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": ["label", "description"],
        "title": "WorkflowStageDisplay",
        "type": "object"
      },
      "WorkspaceListResponse": {
        "description": "工作区切换器使用的游标分页结果。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": ["items"],
        "title": "WorkspaceListResponse",
        "type": "object"
      },
      "WorkspaceResponse": {
        "description": "工作区列表和详情共用的可展示字段。",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "research_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Research Question"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workflow_stage": {
            "$ref": "#/components/schemas/WorkspaceWorkflowStage"
          },
          "workflow_stage_display": {
            "$ref": "#/components/schemas/WorkflowStageDisplay"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "research_question",
          "status",
          "workflow_stage",
          "workflow_stage_display",
          "created_at",
          "updated_at"
        ],
        "title": "WorkspaceResponse",
        "type": "object"
      },
      "WorkspaceWorkflowStage": {
        "description": "工作区在研究主流程中的阶段；英文值会写入数据库和 API。",
        "enum": [
          "draft",
          "analyzing",
          "plan_review",
          "retrieving",
          "screening",
          "collection_building",
          "researching",
          "failed"
        ],
        "title": "WorkspaceWorkflowStage",
        "type": "string"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "面向学术文献检索与 RAG 研究工作区的后端服务。",
    "title": "Academic Search API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/collections": {
      "get": {
        "description": "供工作区切换器按需加载，默认只显示活动工作区。",
        "operationId": "list_workspaces_api_v1_collections_get",
        "parameters": [
          {
            "description": "是否包含已归档工作区",
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "description": "是否包含已归档工作区",
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "description": "按工作区名称或当前研究阶段搜索",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按工作区名称或当前研究阶段搜索",
              "title": "Q"
            }
          },
          {
            "description": "上一页返回的不透明分页游标",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1000,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "上一页返回的不透明分页游标",
              "title": "Cursor"
            }
          },
          {
            "description": "本次最多返回的工作区数量",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "本次最多返回的工作区数量",
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "搜索我的研究工作区",
        "tags": ["研究工作区"]
      },
      "post": {
        "description": "创建一个空工作区；研究问题将在进入工作区后另行确认。",
        "operationId": "create_workspace_api_v1_collections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "创建研究工作区",
        "tags": ["研究工作区"]
      }
    },
...
    "/api/v1/collections/{collection_id}/build": {
      "post": {
        "description": "将所有待确认文献投递到 RAG 入库 Worker；每篇文献独立报告投递结果。",
        "operationId": "build_collection_api_v1_collections__collection_id__build_post",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Collection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionBuildResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "确认并构建研究集合",
        "tags": ["研究集合构建"]
      }
    },
...
    "/api/v1/collections/{collection_id}/documents": {
      "get": {
        "description": "页面刷新时读取活动文献、失败原因与可以进入研究对话的文献数量。",
        "operationId": "list_collection_documents_api_v1_collections__collection_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Collection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDocumentsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "获取集合文献及入库状态",
        "tags": ["研究集合构建"]
      }
    },
    "/api/v1/collections/{collection_id}/documents/{document_id}": {
      "delete": {
        "description": "仅归档 pending 文献，保留正式对象和审计记录以避免跨服务删除不一致。",
        "operationId": "remove_pending_document_api_v1_collections__collection_id__documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Collection Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDocumentRemovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "从待确认集合移出文献",
        "tags": ["研究集合构建"]
      }
    },
    "/api/v1/collections/{collection_id}/ingestion-runs/{ingestion_run_id}/retry": {
      "post": {
        "description": "新建入库运行而非覆盖失败记录，使错误和重试序号始终可审计。",
        "operationId": "retry_ingestion_run_api_v1_collections__collection_id__ingestion_runs__ingestion_run_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Collection Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "ingestion_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ingestion Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionBuildResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "重试失败的文献入库",
        "tags": ["研究集合构建"]
      }
    },
...
  }
}